feat: moved from details/my-projects
