+++ src/projects/presentation/components/detail/my-projects/ProjectProgressCard.vue
